Reorganized files in FE
