--- frontend/src/NewDiary.js
@@ -1,9 +1,0 @@
-import React, { Component } from 'react'
-class NewDiary extends Component {
-    render() {
-      return(
-      <p>New Diary</p>
-    )
-   }
-}
-export default NewDiary;
