--- frontend/src/NewEntry.js
@@ -1,9 +1,0 @@
-import React, { Component } from 'react'
-class NewEntry extends Component {
-    render() {
-      return(
-      <p>New Entry</p>
-    )
-   }
-}
-export default NewEntry;
