--- frontend/src/Welcome.js
@@ -1,9 +1,0 @@
-import React, { Component } from 'react'
-class Welcome extends Component {
-    render() {
-      return(
-      <p>Welcome</p>
-    )
-   }
-}
-export default Welcome;
